Refactor
Remove template files and customise file and folder setup.
