+++ public/index.html
@@ -5,8 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="theme-color" content="#000000">
 
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
+    <!--<link rel="manifest" href="%PUBLIC_URL%/manifest.json">-->
+    <!--<link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">-->
 
     <title>Project Vader</title>
   </head>
