+++ src/components/Core/Core.js
@@ -1,0 +1,11 @@
+import React, {Component, Fragment} from 'react';
+
+export default class Core extends Component {
+  render() {
+    return (
+    <Fragment>
+      <h1>Welcome to Project Vader!</h1>
+    </Fragment>
+    );
+  }
+}
