+++ src/lib/registerServiceWorker.js
