+++ src/styles/index.css
@@ -1,0 +1,4 @@
+body {
+  margin: 0;
+  padding: 0;
+  font-family: sans-serif; }
